pyproject.toml
[tool.poetry]
name = "fastapiproject"
version = "0.1.0"
description = ""
authors = ["Madi <abaymadi1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.115.0"
uvicorn = "0.30.6"
python-multipart = "^0.0.6"
ormar = "0.20.1"
alembic = "1.13.3"
websockets = "^11.0.3"
redis = "^4.5.5"
annotated-types = "0.7.0"
anyio = "4.6.0"
asyncpg = "0.29.0"
click = "8.1.7"
databases = "0.7.0"
h11 = "0.14.0"
idna = "3.10"
mako = "1.3.5"
markupsafe = "2.1.5"
psycopg2-binary = "2.9.9"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
sniffio = "1.3.1"
sqlalchemy = "1.4.54"
starlette = "0.38.6"
typing-extensions = "4.12.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
